refactor: adequando os tipos de dados nos models e validations e limitando caracteres
