+++ backend/src/domain/enderecos/models/Enderecos.js
@@ -1,4 +1,4 @@
-const db = require("../../../infrastructure/database");
+const db = require("../../../infrastructure/database/dbConexao");
 const { DataTypes } = require("sequelize");
 
 const Enderecos = db.define(
@@ -11,31 +11,31 @@ const Enderecos = db.define(
       type: DataTypes.INTEGER,
     },
     rua: {
-      type: DataTypes.STRING,
+      type: DataTypes.STRING(45),
       allowNull: false,
     },
     numero: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.STRING(10),
       allowNull: false,
     },
     bairro: {
-      type: DataTypes.STRING,
+      type: DataTypes.STRING(45),
       allowNull: false,
     },
     cep: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: false,
     },
     complemento: {
-      type: DataTypes.STRING,
+      type: DataTypes.STRING(45),
       allowNull: true,
     },
     cidade: {
-      type: DataTypes.STRING,
+      type: DataTypes.STRING(45),
       allowNull: false,
     },
     estado: {
-        type: DataTypes.STRING,
+        type: DataTypes.STRING(45),
         allowNull: false,
     },
     criadoEm: {
