+++ backend/src/domain/pacientes/validations/atualizar.js
@@ -6,15 +6,14 @@ const validacaoAtualizar = validate({
     }),
     body:Joi.object({
         nome:Joi.string(),
-        cpf:Joi.number(),
+        cpf:Joi.string(),
         data_nascimento:Joi.date(),
         email:Joi.string(),
-        telefone:Joi.number(),
+        telefone:Joi.string(),
         observacoes:Joi.string(),
         avatar:Joi.string(),
-        situacao:Joi.string(),
         rua:Joi.string(),
-        numero:Joi.number(),
+        numero:Joi.string(),
         bairro:Joi.string(),
         cep:Joi.number(),
         complemento:Joi.string(),
