+++ backend/src/domain/pacientes/validations/cadastrar.js
@@ -3,13 +3,12 @@ const {validate, Joi} = require('express-validation')
 const validacaoCadastrar = validate({
     body:Joi.object({
         nome:Joi.string().required(),
-        cpf:Joi.number().required(),
+        cpf:Joi.string().required(),
         data_nascimento:Joi.date().required(),
         email:Joi.string().required(),
-        telefone:Joi.number().required(),
+        telefone:Joi.string().required(),
         observacoes:Joi.string().required(),
         avatar:Joi.string().required(),
-        situacao:Joi.string().required(),
         rua:Joi.string().required(),
         numero:Joi.number().required(),
         bairro:Joi.string().required(),
